Use pydrm to control display
This is a temporary solution that switches pygame SDL output to dummy
device. When a new image is about to be displayed a content of dummy device
is converted to PIL image. The PIL image is displayed using a pydrm library.

Unfortunately, the pydrm cannot be de-initialized. Therefore it is
initialized as static part of Screen object.

The pygame is left in place for compatibility and is still used for
bitmap manipulations and histogram computation.