sl1fw: set path to weston-framebuffer's fb
* python always truncates a file opened for writing: use lower-level
  functions to circumvent that
* prepare a mock fb file of required size during setup phase of
  integration tests
* set up mock fb file for libScreen unittest
* skip setting fb black on initialization (it already is black)
* drop call to fbset, as it becomes meaningless with an emulated fb

Skipping of painting the fb black on init is motivated by the easily
avoidable requirement it puts on the existence of the file early in
a64-fw's initialization. It is undesirable to make a64-fw start only
after weston-framebuffer goes ready as it would prolong the overall
boot-time.
An alternative arrangement could be to relegate the responsibility
of setting up /dev/shm/fb to the a64-fw and have weston-framebuffer
service postponed until a64-fw comes up.