Enable factory mode by file presence
Enable factory mode by presence of the:

/usr/share/factory/defaults/factory_mode_enabled

file. This one will be used to trigger factory mode in a64-fw and also
to trigger ssh and uart console. File presence is used instead of
structured data as systemd can only trigger services based on the file
presence not the content. Old way of factory mode enabling is
kept for compatibility.