Rework printer state hadling, fix admin
As admin can be entered while idle or printing it is not enough to set
state on admin enter and reset od state on leave. The old state might
have changed in the meantime. This change introduces a set of states
entered by the printer. The overall printer state is defined as the most
important state in the set. This way admin is added to the set on enter
and removed on leave. This solves admin while printing problem and
simplifies state change hadling methods.

Fixes SL1SW-1144