Add pour in resin state to exposure
This add new "pour in resin" state between print confirm and checks. The
purpose of the state is to let the user pour in the resin once the tower
is up and tilt leveled. This new state is preceeded by "going up" and new
support state "leveling tilt".

Part of: SL1SW-1033