Fix failed exposure duration
As we set initial exposure timestamp at the start of the exposure itself
the start time is not set during the preprint checks. In case of a
preprint check failure the failed print report includes a time computed
based on the 1.1.1970 timestamp. This sets the initial timestamp value
to now. This produces sane failed report and keeps the old time
computation as the value is reset later on.