Fix vat revision on factory reset
On factory reset set the vat revision = 1 in case of printer being
SL1S as this is what user usually expects. On the other hand, those who
were using metalic vat together with SL1S printer will be forced to
change the settings back in admin as otherwise the resin volume will be
wrong.

Fixes: SL1SW-1347