Add support for fixed point numbers to admin
This new admin item is similar to float, but addresses rounding errors
by using fixed point numbers to represent the value. Internally the
value is passed as an intenger together with another integer specifying
how many decimal digits are dedicated to fractional part of the number.

Part of: SL1SW-1089